rec_system_api/requirements.txt
fastapi
uvicorn
rec-system-trb
mysql-connector-python